+++ src/app/asisten/page.tsx
@@ -7,6 +7,7 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
+import { extractDate, extractTime } from '@/lib/utils';
 import {
   BookOpen,
   Calendar,
@@ -136,26 +137,7 @@ export default async function AsistenPage() {
     </>
   );
 }
-/**
- * Extract date in format YYYY-MM-DD from Date or ISO string
- */
-export function extractDate(date: Date | string): string {
-  const d = typeof date === 'string' ? new Date(date) : date;
-  // timezone handling to +0700
-  d.setHours(d.getHours() + 7);
-  // return date in YYYY-MM-DD format
-  return d.toISOString().slice(0, 10);
-}
 
-/**
- * Extract time in format HH:mm from Date or ISO string
- */
-export function extractTime(date: Date | string): string {
-  const d = typeof date === 'string' ? new Date(date) : date;
-  // timezone handling to +0700
-  d.setHours(d.getHours() + 7);
-  return d.toISOString().slice(11, 16);
-}
 export interface Data {
   jumlah_praktikan_belum_dinilai: number;
   sisa_pertemuan_praktikum: number;
