+++ src/components/layout/asisten/laporan/list/columns.tsx
@@ -18,6 +18,7 @@ import {
 } from 'lucide-react';
 import Link from 'next/link';
 
+import { extractDate, extractTime } from '@/lib/utils';
 import EditFormLaporanButton from '../edit-form';
 
 export interface Laporan {
@@ -33,26 +34,7 @@ export interface Jadwal {
   selesai: Date;
   is_dilaksanakan: boolean;
 }
-/**
- * Extract date in format YYYY-MM-DD from Date or ISO string
- */
-export function extractDate(date: Date | string): string {
-  const d = typeof date === 'string' ? new Date(date) : date;
-  // timezone handling to +0700
-  d.setHours(d.getHours() + 7);
-  // return date in YYYY-MM-DD format
-  return d.toISOString().slice(0, 10);
-}
 
-/**
- * Extract time in format HH:mm from Date or ISO string
- */
-export function extractTime(date: Date | string): string {
-  const d = typeof date === 'string' ? new Date(date) : date;
-  // timezone handling to +0700
-  d.setHours(d.getHours() + 7);
-  return d.toISOString().slice(11, 16);
-}
 export const columns: ColumnDef<Laporan>[] = [
   {
     id: 'select',
