+++ src/components/layout/praktikan/register/register-form.tsx
@@ -46,7 +46,7 @@ import {
   TableHeader,
   TableRow,
 } from '@/components/ui/table';
-import { cn } from '@/lib/utils';
+import { cn, extractDate, extractDay, extractTime } from '@/lib/utils';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useGetCookie } from 'cookies-next/client';
 import {
@@ -556,31 +556,4 @@ export default function RegisterForm() {
       </CardContent>
     </Card>
   );
-}
-
-function extractDay(date: Date | string): string {
-  const d = typeof date === 'string' ? new Date(date) : date;
-  const days = ['Minggu', 'Senin', 'Selasa', 'Rabu', 'Kamis', 'Jumat', 'Sabtu'];
-  return days[d.getDay()];
-}
-
-/**
- * Extract date in format YYYY-MM-DD from Date or ISO string
- */
-export function extractDate(date: Date | string): string {
-  const d = typeof date === 'string' ? new Date(date) : date;
-  // timezone handling to +0700
-  d.setHours(d.getHours() + 7);
-  // return date in YYYY-MM-DD format
-  return d.toISOString().slice(0, 10);
-}
-
-/**
- * Extract time in format HH:mm from Date or ISO string
- */
-export function extractTime(date: Date | string): string {
-  const d = typeof date === 'string' ? new Date(date) : date;
-  // timezone handling to +0700
-  d.setHours(d.getHours() + 7);
-  return d.toISOString().slice(11, 16);
 }
